+++ .github/workflows/test.yml
@@ -20,7 +20,7 @@ jobs:
     steps:
       - name: Install Poetry
         uses: snok/install-poetry@v1
-      - name: Checkout mailgunsdk-sdk
+      - name: Checkout mailgun_sdk-sdk
         uses: actions/checkout@v2
 
       - name: Set up Python
