+++ mailgun_sdk/base.py
