+++ tests/test_init.py
@@ -1,13 +1,13 @@
 import unittest
 
-import mailgunsdk
-from mailgunsdk.api import MailgunApi
+import mailgun_sdk
+from mailgun_sdk.api import MailgunApi
 
 
 class MailgunInitTestCase(unittest.TestCase):
     def test_api(self):
-        self.assertIsInstance(mailgunsdk.api, MailgunApi)
+        self.assertIsInstance(mailgun_sdk.api, MailgunApi)
 
     def test_initialize(self):
-        mailgunsdk.initialize("api-key-xxx")
-        self.assertEqual(mailgunsdk.api.api_key, "api-key-xxx")
+        mailgun_sdk.initialize("api-key-xxx")
+        self.assertEqual(mailgun_sdk.api.api_key, "api-key-xxx")
